Saving

+++ FinancialMachineLearning/barsampling/base_bars.py
@@ -204,7 +204,7 @@ class BaseImbalanceBars(BaseBars):
             # Set variables
             date_time = row[0]
             self.tick_num += 1
-            price = np.float(row[1])
+            price = np.float64(row[1])
             volume = row[2]
             dollar_value = price * volume
             signed_tick = self._apply_tick_rule(price)
@@ -299,7 +299,7 @@ class BaseRunBars(BaseBars):
         for row in data:
             date_time = row[0]
             self.tick_num += 1
-            price = np.float(row[1])
+            price = np.float64(row[1])
             volume = row[2]
             dollar_value = price * volume
             signed_tick = self._apply_tick_rule(price)
